LSM: Create SetAssociativeCache (WIP, happy holidays)
The beginning of a set-associative cache for caching billions of values,
with an Nth-chance CLOCK replacement policy for very low miss rates.

This is a work-in-progress at a VERY rough early stage, more to come!

The layout is configurable to benchmark and evaluate different layouts.

The big idea is to separate the values (8-65536 bytes), the tags
(1-2 bytes per value), the CLOCK counters (1-2 bits per value),
and the CLOCK hands (2-4 bits per set of 2,4 or 16 values).

This means we can then eliminate the padding that would otherwise be
introduced by the 2-4 bit CLOCK hands per set, so that we can keep
everything cache-line aligned and also improve the odds that most
metadata (counters, hands) will reside in the CPU cache.

For example, a 64 GiB 16-way cache for 64 KiB blocks, with 8-bit tags
and 2-bit CLOCK counters would have these contiguous allocations:

1048576 x 64 KiB blocks                       =  64 GiB
-------------------------------------------------------
1048576 x 8-bit SIMD tags                     =   1 MiB
1048576 x 2-bit CLOCK counters                = 256 KiB
1048576 / 16 ways * 4-bit CLOCK hands per set =  32 KiB

The key insight is that most of the time, we're only working in the
metadata zone, comparing against tags and checking CLOCK counters
(using the CLOCK counter also as a valid bit), and only very occasionally
(if the SIMD tag happens to match) having to do a cache miss into the
large blocks zone to compare against the actual key stored in the value.

The reason we don't support an 8-way set-associative cache is because the
CLOCK hands per set would then no longer be cache-line aligned. 16-way is
also a sweet spot for 128-bit SSE SIMD, which is widely supported.

Under the hood, this uses a shiny new PackedUnsignedIntegerArray to pack
the small (less than 8-bit) CLOCK counters and hands into 64-bit words,
without padding, little-endian only.

Happy holidays!